+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add :bar_format option to allow selecting preconfigured bar displays
 * Add :eta_time format token to display the estimated time of day at completion
 * Add measurement of the total elapsed time that ignores stopped time intervals
+* Add #pause to prevent bar from continuing progression and suspend time measurements
 
 ### Changed
 * Change Multi#stopped? to check that all bars are stopped